add authorService tests

+++ services/authorsService.ts
@@ -44,7 +44,7 @@ const getById = async (authorId: string) => {
 };
 
 const updateById = async (authorId: string, author: AuthorDto) => {
-  const updatedAuthor = await Author.findByIdAndUpdate(authorId, author);
+  const updatedAuthor = await Author.findByIdAndUpdate(authorId, author, {new:true});
   if (!author) {
     throw ApiError.resourceNotFound('Author not found');
   }
@@ -52,6 +52,10 @@ const updateById = async (authorId: string, author: AuthorDto) => {
 };
 
 const deleteById = async (authorId: string) => {
+  const author = await Author.findById(authorId);
+  if (!author) {
+    throw ApiError.resourceNotFound('Author not found');
+  }
   await Author.findByIdAndDelete(authorId);
   await BookAuthor.deleteMany({ authorId });
 };
